warning--
SVN revision: 55476

+++ legacy/eina/src/lib/eina_file_win32.c
@@ -66,7 +66,7 @@ struct _Eina_File_Iterator
 
    WIN32_FIND_DATA data;
    HANDLE          handle;
-   int             length;
+   size_t          length;
    Eina_Bool       is_last : 1;
 
    char            dir[1];
@@ -78,7 +78,7 @@ struct _Eina_File_Direct_Iterator
 
    WIN32_FIND_DATA       data;
    HANDLE                handle;
-   int                   length;
+   size_t                length;
    Eina_Bool             is_last : 1;
 
    Eina_File_Direct_Info info;
